Semifinal implementation of selectable scales
This commit contains a functional implementation of regionally-defined scales. This feature allows regional admins to define different scales for their region, which are named and associated with a unit of weight. This commit does not contain a functional unit conversion mechanism - the code is mostly in place, but with the actual conversion dummied out because of some bugs I have yet to tackle.